+++ src/Math.cpp
@@ -230,21 +230,21 @@ namespace kann::math
 
   Tensor add(Tensor a, Tensor b)
   {
-    return cwise(std::move(a), std::move(b), [](double a, double b){
+    return cwise(std::move(a), std::move(b), [](float a, float b){
       return a+b;
     });
   }
 
   Tensor sub(Tensor a, Tensor b)
   {
-    return cwise(std::move(a), std::move(b), [](double a, double b){
+    return cwise(std::move(a), std::move(b), [](float a, float b){
       return a-b;
     });
   }
 
   Tensor scale(Tensor a, float factor)
   {
-    return cwise(std::move(a), [&](double a){
+    return cwise(std::move(a), [&](float a){
       return a * factor;
     });
   }
